# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portfolio",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://http://localhost:82/swagger/index.html",
            "description": "Mon Api pour portfolio"
        }
    ],
    "paths": {
        "/activities/{id}": {
            "get": {
                "tags": [
                    "Secteur d'activité"
                ],
                "operationId": "App\\Controller\\ActivityAreaController::showAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Secteur d'activité",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "put": {
                "tags": [
                    "Secteur d'activité"
                ],
                "operationId": "App\\Controller\\ActivityAreaController::putAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Secteur d'activité",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "post": {
                "tags": [
                    "Secteur d'activité"
                ],
                "operationId": "App\\Controller\\ActivityAreaController::createAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateActivityArea"
                },
                "responses": {
                    "200": {
                        "description": "Notre Secteur d'activité",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Secteur d'activité"
                ],
                "operationId": "App\\Controller\\ActivityAreaController::removeActivityAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Secteur d'activité",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Secteur d'activité"
                ],
                "operationId": "App\\Controller\\ActivityAreaController::patchAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Secteur d'activité",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            }
        },
        "/activities/": {
            "get": {
                "tags": [
                    "Liste des Secteurs d'activité"
                ],
                "operationId": "App\\Controller\\ActivityAreaController::listAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résume des Secteurs d'activité",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityQuickView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/candidats/{id}": {
            "get": {
                "tags": [
                    "Candidat"
                ],
                "operationId": "App\\Controller\\CandidatController::showAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Candidat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Candidat"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "put": {
                "tags": [
                    "Candidat"
                ],
                "operationId": "App\\Controller\\CandidatController::putAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Candidat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Candidat"
                                }
                            }
                        }
                    },
                    "204": {
                        "$ref": "#/components/responses/204 -  NO CONTENT"
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidat"
                ],
                "operationId": "App\\Controller\\CandidatController::createAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateCandidat"
                },
                "responses": {
                    "201": {
                        "description": "Notre Candidat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Candidat"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    },
                    "409": {
                        "$ref": "#/components/responses/409 - CONFLICT"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Candidat"
                ],
                "operationId": "App\\Controller\\CandidatController::removeCandidatAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Candidat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Candidat"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Candidat"
                ],
                "operationId": "App\\Controller\\CandidatController::patchAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Candidat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Candidat"
                                }
                            }
                        }
                    },
                    "204": {
                        "$ref": "#/components/responses/204 -  NO CONTENT"
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            }
        },
        "/candidats/": {
            "get": {
                "tags": [
                    "Liste des Candidat"
                ],
                "operationId": "App\\Controller\\CandidatController::listAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Liste de Candidat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CandidatQuickView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/companies/{id}": {
            "get": {
                "tags": [
                    "Entreprise"
                ],
                "operationId": "App\\Controller\\CompanyController::showAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Entreprise",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "put": {
                "tags": [
                    "Entreprise"
                ],
                "operationId": "App\\Controller\\CompanyController::putAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "post": {
                "tags": [
                    "Entreprise"
                ],
                "operationId": "App\\Controller\\CompanyController::createAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateCompany"
                },
                "responses": {
                    "200": {
                        "description": "Notre Company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Entreprise"
                ],
                "operationId": "App\\Controller\\CompanyController::removeCompanyAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Entreprise"
                ],
                "operationId": "App\\Controller\\CompanyController::patchAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            }
        },
        "/companies/": {
            "get": {
                "tags": [
                    "Liste des Entreprises"
                ],
                "operationId": "App\\Controller\\CompanyController::listAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Liste des Entreprises",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyQuickView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/diplomes/{id}": {
            "get": {
                "tags": [
                    "Diplome"
                ],
                "operationId": "App\\Controller\\DiplomeController::showAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Diplome",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Diplome"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "put": {
                "tags": [
                    "Diplome"
                ],
                "operationId": "App\\Controller\\DiplomeController::createAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Diplome",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Diplome"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "post": {
                "tags": [
                    "Diplome"
                ],
                "operationId": "App\\Controller\\DiplomeController::createAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateDiplome"
                },
                "responses": {
                    "200": {
                        "description": "Notre Diplome",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Diplome"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Diplome"
                ],
                "operationId": "App\\Controller\\DiplomeController::removeDiplomeAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Diplome",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Diplome"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Diplome"
                ],
                "operationId": "App\\Controller\\DiplomeController::createAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notre Diplome",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Diplome"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404 - NotFound"
                    }
                }
            }
        },
...
        "/formations/": {
            "get": {
                "tags": [
                    "Liste des Formations"
                ],
                "operationId": "App\\Controller\\FormationController::listAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des Formations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FormationQuickView"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/langues/": {
            "get": {
                "tags": [
                    "Liste des langues"
                ],
                "operationId": "App\\Controller\\LangueController::listAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des langues",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LangueQuickView"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/licenses/": {
            "get": {
                "tags": [
                    "Liste des licenses"
                ],
                "operationId": "App\\Controller\\LicenseController::listAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des Licenses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LicenseQuickView"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/schools/": {
            "get": {
                "tags": [
                    "Liste des Ecoles"
                ],
                "operationId": "App\\Controller\\SchoolController::listAction",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des Ecoles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SchoolQuickView"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ActivityQuickView": {
                "description": "Résume des Secteurs d'activité",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": "true"
                    }
                },
                "type": "object"
            },
            "Activity": {
                "description": "Notre Secteur d'activité",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": "true"
                    }
                },
                "type": "object"
            },
            "CandidatQuickView": {
                "description": "Notre Candidat en résumé",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "firstname": {
                        "type": "string",
                        "nullable": "true"
                    },
                    "lastname": {
                        "type": "string",
                        "nullable": "true"
                    },
                    "short_description": {
                        "type": "string",
                        "nullable": "true"
                    }
                },
                "type": "object"
            },
...
            "Company": {
                "description": "Notre Entreprise",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": "true"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": "true"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": "true"
                    },
                    "activityAreas": {
                        "items": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/Activity"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "ID du candidat",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "requestBodies": {
            "UpdateActivityArea": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {}
                    }
                }
            },
            "UpdateCandidat": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {}
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    }
}